#123: - reorganize structure, so app is under `src`
