+++ src/Execution/lbsTemperedCriterion.py
@@ -48,7 +48,7 @@ from src.Execution.lbsCriterionBase import CriterionBase
 from src.Model.lbsObject import Object
 from src.Model.lbsObjectCommunicator import ObjectCommunicator
 from src.Model.lbsRank import Rank
-from utils.logger import CLRS
+from src.Utils.logger import CLRS
 
 
 class TemperedCriterion(CriterionBase):
