+++ src/Utils/logger.py
@@ -2,8 +2,8 @@ import logging
 import logging.config
 import os
 
-from utils.colors import red, green, blue, cyan, magenta, yellow, white, black, light_red, light_green, light_blue, \
-    light_cyan, light_magenta, light_yellow, light_white, light_black
+from src.Utils.colors import red, green, blue, cyan, magenta, yellow, white, black, light_red, light_green, \
+    light_blue, light_cyan, light_magenta, light_yellow, light_white, light_black
 
 _cfg_path = os.path.join(os.path.dirname(__file__), 'logger.ini')
 logging.config.fileConfig(_cfg_path)
